requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
